+++ src/entities/video-contents/apis/index.ts
@@ -1,3 +1,4 @@
 export * from "./getVideoContentsApi";
 export * from "./videoContentsApi";
 export * from "./getSeriesTitleApi";
+export * from "./updateVideoApi";
